attendee handle exception

+++ src/main/java/com/example/springboot03/service/serviceImpl/AttendeeServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.springboot03.service.serviceImpl;
 
+import com.example.springboot03.exception.NotFoundException;
 import com.example.springboot03.model.dto.request.AttendeeRequest;
 import com.example.springboot03.model.entity.Attendee;
 import com.example.springboot03.repository.AttendeeRepository;
@@ -11,6 +12,7 @@ import java.util.List;
 @Service
 public class AttendeeServiceImpl implements AttendeeService {
     final AttendeeRepository attendeeRepository;
+
     public AttendeeServiceImpl(AttendeeRepository attendeeRepository) {
         this.attendeeRepository = attendeeRepository;
     }
@@ -18,26 +20,41 @@ public class AttendeeServiceImpl implements AttendeeService {
     @Override
     public List<Attendee> getAllAttendees(Integer offset, Integer limit) {
         Integer page = (offset - 1) * limit;
-        return attendeeRepository.getAllAttendee(page, limit);
+        return attendeeRepository.getAllAttendees(page, limit);
     }
 
     @Override
     public Attendee addAttendee(AttendeeRequest attendeeRequest) {
+        if (attendeeRequest == null || attendeeRequest.getName().isEmpty()) {
+            throw new BadRequestException("Invalid attendee data");
+        }
         return attendeeRepository.addAttendee(attendeeRequest);
     }
 
     @Override
     public Attendee getAttendeeById(Integer id) {
-        return attendeeRepository.getAttendeeById(id);
+        Attendee attendee = attendeeRepository.getAttendeeById(id);
+        if (attendee == null) {
+            throw new NotFoundException("Attendee not found with ID: " + id);
+        }
+        return attendee;
     }
 
     @Override
     public Attendee updateAttendeeById(Integer id, AttendeeRequest attendeeRequest) {
+        Attendee attendee = attendeeRepository.getAttendeeById(id);
+        if (attendee == null) {
+            throw new NotFoundException("Attendee not found with ID: " + id);
+        }
         return attendeeRepository.updateAttendeeById(id, attendeeRequest);
     }
 
     @Override
     public Attendee deleteAttendeeById(Integer id) {
+        Attendee attendee = attendeeRepository.getAttendeeById(id);
+        if (attendee == null) {
+            throw new NotFoundException("Attendee not found with ID: " + id);
+        }
         return attendeeRepository.deleteAttendeeById(id);
     }
 }
